+++ src/generators/dom/visitors/Element/Binding.ts
@@ -1,6 +1,5 @@
 import deindent from '../../../../utils/deindent';
 import flattenReference from '../../../../utils/flattenReference';
-import getSetter from '../shared/binding/getSetter';
 import getStaticAttributeValue from './getStaticAttributeValue';
 import { DomGenerator } from '../../index';
 import Block from '../../Block';
@@ -50,16 +49,7 @@ export default function visitBinding(
 		type
 	);
 
-	let setter = getSetter({
-		block,
-		name,
-		snippet,
-		_this: state.parentNode,
-		props: '_svelte',
-		attribute,
-		dependencies,
-		value,
-	});
+	let setter = getSetter(block, name, snippet, state.parentNode, attribute, dependencies, value);
 	let updateElement = `${state.parentNode}.${attribute.name} = ${snippet};`;
 	const lock = `#${state.parentNode}_updating`;
 	let updateCondition = `!${lock}`;
@@ -270,4 +260,59 @@ function getBindingGroup(generator: DomGenerator, value: Node) {
 	}
 
 	return index;
+}
+
+function getSetter(
+	block: Block,
+	name: string,
+	snippet: string,
+	_this: string,
+	attribute: Node,
+	dependencies: string[],
+	value: string,
+) {
+	const tail = attribute.value.type === 'MemberExpression'
+		? getTailSnippet(attribute.value)
+		: '';
+
+	if (block.contexts.has(name)) {
+		const prop = dependencies[0];
+		const computed = isComputed(attribute.value);
+
+		return deindent`
+			var list = ${_this}._svelte.${block.listNames.get(name)};
+			var index = ${_this}._svelte.${block.indexNames.get(name)};
+			${computed && `var state = #component.get();`}
+			list[index]${tail} = ${value};
+
+			${computed
+				? `#component.set({ ${dependencies
+						.map((prop: string) => `${prop}: state.${prop}`)
+						.join(', ')} });`
+				: `#component.set({ ${dependencies
+						.map((prop: string) => `${prop}: #component.get( '${prop}' )`)
+						.join(', ')} });`}
+		`;
+	}
+
+	if (attribute.value.type === 'MemberExpression') {
+		return deindent`
+			var state = #component.get();
+			${snippet} = ${value};
+			#component.set({ ${dependencies
+				.map((prop: string) => `${prop}: state.${prop}`)
+				.join(', ')} });
+		`;
+	}
+
+	return `#component.set({ ${name}: ${value} });`;
+}
+
+function isComputed(node: Node) {
+	while (node.type === 'MemberExpression') {
+		if (node.computed) return true;
+		node = node.object;
+	}
+
+	return false;
 }
