+++ test/js/samples/collapses-text-around-comments/expected.js
@@ -65,6 +65,7 @@ function SvelteComponent ( options ) {
 
 	this._root = options._root || this;
 	this._yield = options._yield;
+	this._bind = options._bind;
 
 	if ( !document.getElementById( 'svelte-3590263702-style' ) ) add_css();
