+++ test/js/samples/computed-collapsed-if/expected.js
@@ -38,6 +38,7 @@ function SvelteComponent ( options ) {
 
 	this._root = options._root || this;
 	this._yield = options._yield;
+	this._bind = options._bind;
 
 	this._fragment = create_main_fragment( this._state, this );
