+++ test/js/samples/each-block-changed-check/expected.js
@@ -152,6 +152,7 @@ function SvelteComponent ( options ) {
 
 	this._root = options._root || this;
 	this._yield = options._yield;
+	this._bind = options._bind;
 
 	this._fragment = create_main_fragment( this._state, this );
