+++ test/js/samples/if-block-simple/expected.js
@@ -76,6 +76,7 @@ function SvelteComponent ( options ) {
 
 	this._root = options._root || this;
 	this._yield = options._yield;
+	this._bind = options._bind;
 
 	this._fragment = create_main_fragment( this._state, this );
